[NFC] Outline an express app into a separate module
This is done in order to be able to re-use that app in API unit tests later
